docs & style: Put docs in ./src directory and rename extension
The source of the documentatation shall be contained in `./docs/src`. These
are the files that are processed into the official documentation.

Templates, static assets, and other helpful things used in the official docs
can be added under the `./docs` folder as well.

Breaking: Github will no longer render these as Markdown since the extensions
are now `.docfu` (the documentation generator).